requirements.txt
pyyaml
dnspython
psutil
boto3>=1.4.1
requests
